+++ CHANGELOG.md
@@ -1,25 +1,36 @@
 # Changelog
 
-## v1.1.2 - 2021-06-08
+## [UNRELEASED]
 
-* Fix issue at startup with gpio configured
+### Changed
+* Frontend service doesn't have to catch update events
+
+## [1.1.2] - 2021-06-08
+
+### Changed
 * Improve command parameters check
 
-## v1.1.1 - 2021-05-16
+### Fixed
+* Fix issue at startup with gpio configured
 
+## [1.1.1] - 2021-05-16
+
+### Changed
 * Backend: migrate to python3
 * Backend: update after core changes
 * Backend: improve unit tests + fix small issues
 * Frontend: code update
 
-## v1.1.0 - 2021-04-08
+## [1.1.0] - 2021-04-08
 
+### Changed
 * Improve robustness adding unit tests
 * Add documentation generation
 * Update after core changes
 * Fox some issues
 
-## v1.0.0 - 2019-02-16
+## [1.0.0] - 2019-02-16
 
+### Changed
 * First release